Revert "PR for #4406 - Add ability to enable/disable authentication for the MicroProfile Metrics endpoint"
