+++ dev/com.ibm.websphere.appserver.features/visibility/public/mpMetrics-1.0/com.ibm.websphere.appserver.mpMetrics-1.0.feature
@@ -12,9 +12,6 @@ Subsystem-Name: MicroProfile Metrics 1.0
  com.ibm.websphere.appserver.restHandler-1.0, \
  com.ibm.websphere.appserver.servlet-3.1; ibm.tolerates:=4.0
 -bundles=com.ibm.ws.require.java8, \
- com.ibm.ws.microprofile.metrics, \
- com.ibm.ws.microprofile.metrics.1.0, \
- com.ibm.ws.microprofile.metrics.private, \
- com.ibm.ws.microprofile.metrics.public
+ com.ibm.ws.microprofile.metrics
 kind=ga
 edition=core
