+++ dev/com.ibm.websphere.appserver.features/visibility/public/mpMetrics-1.1/com.ibm.websphere.appserver.mpMetrics-1.1.feature
@@ -13,9 +13,6 @@ Subsystem-Name: MicroProfile Metrics 1.1
  com.ibm.websphere.appserver.servlet-3.1; ibm.tolerates:=4.0,\
  com.ibm.websphere.appserver.mpConfig-1.2; ibm.tolerates:=1.3
 -bundles=com.ibm.ws.require.java8, \
- com.ibm.ws.microprofile.metrics, \
- com.ibm.ws.microprofile.metrics.1.1, \
- com.ibm.ws.microprofile.metrics.private, \
- com.ibm.ws.microprofile.metrics.public
+ com.ibm.ws.microprofile.metrics.1.1
 kind=ga
 edition=core
