--- dev/com.ibm.ws.microprofile.metrics.1.0/bnd.bnd
@@ -1,43 +1,0 @@
-#*******************************************************************************
-# Copyright (c) 2018 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License v1.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-v10.html
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-#*******************************************************************************
--include= ~../cnf/resources/bnd/bundle.props
-bVersion=1.0
-
-javac.source: 1.8
-javac.target: 1.8
-
-Bundle-Name: MicroProfile Metrics 1.0
-Bundle-SymbolicName: com.ibm.ws.microprofile.metrics.1.0
-
-Bundle-Description: MicroProfile Metrics 1.0, version ${bVersion}
-
-IBM-Authorization-Roles: com.ibm.ws.management
-IBM-Web-Extension-Processing-Disabled: true
-
-Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
-
--dsannotations-inherit: true
--dsannotations: \
-	com.ibm.ws.microprofile.metrics10.impl.SharedMetricRegistries10
-
-WS-TraceGroup: METRICS
-
--buildpath: \
-	com.ibm.websphere.org.osgi.core;version=latest,\
-	com.ibm.ws.logging;version=latest,\
-	com.ibm.websphere.org.osgi.service.component;version=latest,\
-	com.ibm.wsspi.org.osgi.service.component.annotations;version=latest,\
-	com.ibm.wsspi.org.osgi.service.event;version=latest,\
-	com.ibm.websphere.javaee.annotation.1.2;version=latest,\
-	com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
-	com.ibm.websphere.javaee.interceptor.1.2;version=latest,\
-	com.ibm.websphere.javaee.cdi.1.2;version=latest,\
-	com.ibm.ws.microprofile.metrics;version=latest
