--- dev/com.ibm.ws.microprofile.metrics.1.0/src/com/ibm/ws/microprofile/metrics10/impl/package-info.java
@@ -1,15 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2018 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- * IBM Corporation - initial API and implementation
- *******************************************************************************/
-@org.osgi.annotation.versioning.Version("1.0")
-@TraceOptions(traceGroup = "METRICS", messageBundle = "com.ibm.ws.microprofile.metrics.resources.Metrics")
-package com.ibm.ws.microprofile.metrics10.impl;
-
-import com.ibm.websphere.ras.annotation.TraceOptions;
