+++ dev/com.ibm.ws.microprofile.metrics.1.1/bnd.bnd
@@ -22,6 +22,7 @@ Bundle-Description: MicroProfile Metrics 1.1, version ${bVersion}
 IBM-Authorization-Roles: com.ibm.ws.management
 IBM-Web-Extension-Processing-Disabled: true
 
+Web-ContextPath: /metrics
 
 Import-Package: \
 	javax.enterprise.inject; version="[1.1,3)",\
@@ -38,14 +39,17 @@ Export-Package: \
 	!com.ibm.ws.microprofile.metrics.resources,\
 	com.ibm.ws.microprofile.metrics11.*,\
 	com.ibm.ws.microprofile.metrics.impl,\
-	com.ibm.ws.microprofile.metrics,\
-	com.ibm.ws.microprofile.metrics.helper
+	com.ibm.ws.microprofile.metrics
+
+Include-Resource: \
+    WEB-INF=resources/WEB-INF
     
 Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 
 -dsannotations-inherit: true
--dsannotations: \
-	com.ibm.ws.microprofile.metrics11.impl.SharedMetricRegistries11
+-dsannotations: com.ibm.ws.microprofile.metrics11.impl.SharedMetricRegistries11,\
+	com.ibm.ws.microprofile.metrics.MetricsHandler,\
+	com.ibm.ws.microprofile.metrics.ApplicationListener
 
 WS-TraceGroup: METRICS
