--- dev/com.ibm.ws.microprofile.metrics.private/build.gradle
@@ -1,6 +1,0 @@
-plugins {
-  id "net.ossindex.audit" version "0.3.8-beta"
-  }
-audit {
-    failOnError = false
-}
