+++ dev/com.ibm.ws.microprofile.metrics/bnd.bnd
@@ -14,13 +14,15 @@ bVersion=1.0
 javac.source: 1.8
 javac.target: 1.8
 
-Bundle-Name: MicroProfile Metrics Common
+Bundle-Name: MicroProfile Metrics
 Bundle-SymbolicName: com.ibm.ws.microprofile.metrics
 
-Bundle-Description: MicroProfile Metrics Common, version ${bVersion}
+Bundle-Description: MicroProfile Metrics 1.0, version ${bVersion}
 
 IBM-Authorization-Roles: com.ibm.ws.management
 IBM-Web-Extension-Processing-Disabled: true
+
+Web-ContextPath: /metrics
 
 Import-Package: \
 	javax.enterprise.inject; version="[1.1,3)",\
@@ -31,21 +33,19 @@ Private-Package: \
 	com.ibm.ws.microprofile.metrics.exceptions,\
 	com.ibm.ws.microprofile.metrics.helper,\
 	com.ibm.ws.microprofile.metrics.writer
-
 Export-Package: \
 	com.ibm.ws.microprofile.metrics,\
-	com.ibm.ws.microprofile.metrics.impl,\
-	com.ibm.ws.microprofile.metrics.helper
+	com.ibm.ws.microprofile.metrics.impl
 
 Include-Resource: \
-	OSGI-INF=resources/OSGI-INF
+    WEB-INF=resources/WEB-INF
     
 Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 
 -dsannotations-inherit: true
--dsannotations: \
-	com.ibm.ws.microprofile.metrics.MetricsConfig,\
-	com.ibm.ws.microprofile.metrics.ApplicationListener
+-dsannotations: com.ibm.ws.microprofile.metrics.MetricsHandler,\
+	com.ibm.ws.microprofile.metrics.ApplicationListener,\
+	com.ibm.ws.microprofile.metrics.impl.SharedMetricRegistries
 
 WS-TraceGroup: METRICS
 
@@ -62,7 +62,6 @@ WS-TraceGroup: METRICS
 	com.ibm.websphere.appserver.spi.ssl;version=latest,\
 	com.ibm.wsspi.org.osgi.service.event;version=latest,\
 	com.ibm.websphere.rest.handler;version=latest,\
-	com.ibm.ws.app.manager.wab;version=latest,\
 	com.ibm.ws.rest.handler;version=latest,\
 	com.ibm.ws.webcontainer;version=latest,\
 	com.ibm.ws.container.service;version=latest,\
