--- dev/com.ibm.ws.microprofile.metrics/resources/OSGI-INF/metatype/metatype.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    Copyright (c) 2018 IBM Corporation and others.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-   
-    Contributors:
-        IBM Corporation - initial API and implementation
- -->
-
-<metatype:MetaData xmlns:metatype="http://www.osgi.org/xmlns/metatype/v1.1.0"  
-                   xmlns:ibm="http://www.ibm.com/xmlns/appservers/osgi/metatype/v1.0.0"
-                   localization="OSGI-INF/l10n/metatype">
-
-    <OCD id="com.ibm.ws.microprofile.metrics" name="%mpMetrics" description="%mpMetrics.desc" ibm:alias="mpMetrics">
-        <AD id="authentication" name="%authentication" description="%authentication.desc" required="false" default="true" type="Boolean" />
-    </OCD>
-
-    <Designate pid="com.ibm.ws.microprofile.metrics">
-        <Object ocdref="com.ibm.ws.microprofile.metrics" />
-    </Designate>
-  
-</metatype:MetaData>
