+++ dev/com.ibm.ws.microprofile.metrics/resources/com/ibm/ws/microprofile/metrics/resources/ibm-web-ext.xml
